feat: unify commission components and move data to json
