+++ pages/commission.mdx
@@ -6,9 +6,7 @@ description: The collection of illustrations commissioned by Aozaki.
 image: https://img.aozaki.cc/commission/20211018_Laxy.jpg
 ---
 
-import List from '#commission/list'
-import Featured from '#commission/featured'
-import Stale from '#commission/stale'
+import { Featured, List, Stale } from '#commission/combined'
 
 ## Featured
